添加点赞、修改teamvar sql操作解析等
1. 添加点赞
2. 修改teamvar sql操作解析
3. 添加搜索project, 支持分页
4. 获取当前组织下所有可见project，支持分页
5. 修改部分配置
6. 修改bug

+++ server/src/main/java/edp/davinci/service/impl/ViewServiceImpl.java
@@ -60,6 +60,9 @@ public class ViewServiceImpl extends CommonService<View> implements ViewService 
     private RelUserTeamMapper relUserTeamMapper;
 
     @Autowired
+    private WidgetMapper widgetMapper;
+
+    @Autowired
     private RelTeamProjectMapper relTeamProjectMapper;
 
     @Autowired
@@ -296,6 +299,11 @@ public class ViewServiceImpl extends CommonService<View> implements ViewService 
         if (!allowDelete(viewWithProjectAndSource.getProject(), user)) {
             log.info("user {} have not permisson to delete the view {}", user.getUsername(), id);
             return resultMap.failAndRefreshToken(request, HttpCodeEnum.UNAUTHORIZED).message("you have not permission to delete the view");
+        }
+
+        List<Widget> widgets = widgetMapper.getWidgetsByWiew(id);
+        if (null != widgets && widgets.size() > 0) {
+            return resultMap.failAndRefreshToken(request).message("The current view has been referenced, please delete the reference and then operate");
         }
 
         viewMapper.deleteById(id);
